feat: Add PredictedCapacityGraph and CapacityComparisonGraph components
