+++ frontend/occupi-web/src/components/index.ts
@@ -31,6 +31,9 @@ import ProfileDropDown from "./profileDropDown/ProfileDropDown";
 import AiDashCard from "./aiDashboard/aiDashCard/AiDashCard";
 import RoomsComponent from "./roomsComponent/RoomsComponent";
 import FeedBackModal from "./feedbackModal/FeedBackModal";
+import PredictedCapacityGraph from "./aiDashboard/aiDashGraphs/PredictedCapacityGraph";
+import CapacityComparisonGraph from "./aiDashboard/aiDashGraphs/CapacityComparisonGraph";
+
 
 export {
     DrawerComponent,
@@ -65,5 +68,7 @@ export {
     ProfileDropDown,
     AiDashCard,
     RoomsComponent,
-    FeedBackModal
+    FeedBackModal,
+    PredictedCapacityGraph,
+    CapacityComparisonGraph
 }
